+++ BxInputController/Sources/Rows/Text/View/BxInputChildSelectorTextRowBinder.swift
@@ -45,7 +45,7 @@ open class BxInputChildSelectorTextRowBinder<Row: BxInputChildSelectorTextRow, C
         }
         cell.textView.isUserInteractionEnabled = value
         // UI part
-        if needChangeDisadledCell {
+        if needChangeDisabledCell {
             if let changeViewEnableHandler = owner?.settings.changeViewEnableHandler {
                 changeViewEnableHandler(cell.textView, isEnabled)
             } else {